requirements.txt
searchtweets==1.5.0
twython==3.8.2
newsapi-python==0.2.6